+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/base-journey/services/participant-model-mapper.ts
@@ -1,16 +1,5 @@
 import { HearingSuitabilityAnswer } from 'src/app/services/clients/api-client';
-import { SuitabilityAnswer, HasAccessToCamera, SelfTestAnswers } from '../participant-suitability.model';
-
-export const ParticipantQuestionKeys = {
-  Camera: 'CAMERA_MICROPHONE'
-};
-
-export const QuestionKeys = {
-  Barrister: 'APPOINTING_BARRISTER',
-  BarristerName: 'BARRISTER_NAME',
-  BarristerChambers: 'BARRISTER_CHAMBERS',
-  BarristerEmail: 'BARRISTER_EMAIL'
-};
+import { SuitabilityAnswer, SelfTestAnswers } from '../participant-suitability.model';
 
 export const SelfTestQuestionKeys = {
   CheckYourComputer: 'KIT_CHECK_YOUR_COMPUTER',
@@ -58,23 +47,6 @@ export abstract class ParticipantModelMapper {
     return undefined;
   }
 
-  public mapComputerCamera(answers: HearingSuitabilityAnswer[]): HasAccessToCamera {
-    const answer = answers.find(a => a.question_key === ParticipantQuestionKeys.Camera);
-    if (answer) {
-      switch (answer.answer) {
-        case 'Yes':
-          return HasAccessToCamera.Yes;
-        case 'No':
-          return HasAccessToCamera.No;
-        case 'Not sure':
-          return HasAccessToCamera.NotSure;
-        default:
-          throw new Error(`Unexpected answer to computer question: ${answer.answer}`);
-      }
-    }
-    return undefined;
-  }
-
   public mapBooleanAnswerFromKey(key: string, answers: HearingSuitabilityAnswer[]): SuitabilityAnswer {
     const answer = answers.find(a => a.question_key === key);
     if (answer) {
@@ -108,15 +80,6 @@ export abstract class ParticipantModelMapper {
     }
   }
 
-  public addAnswerForCamera(modelAnswer: HasAccessToCamera, key: string, answers: HearingSuitabilityAnswer[]) {
-    if (modelAnswer !== undefined) {
-      const answer = new HearingSuitabilityAnswer();
-      answer.question_key = key;
-      answer.answer = this.getAccessToCameraAnswer(modelAnswer);
-      answers.push(answer);
-    }
-  }
-
   private createHearingSuitabilityAnswer(key: string, answer: boolean, extendedAnswer: string): HearingSuitabilityAnswer {
     const hearingSuitabilityAnswer = new HearingSuitabilityAnswer();
     hearingSuitabilityAnswer.question_key = key;
@@ -133,18 +96,5 @@ export abstract class ParticipantModelMapper {
     hearingSuitabilityAnswer.answer = answer;
 
     return hearingSuitabilityAnswer;
-  }
-
-  private getAccessToCameraAnswer(accessToCamera: HasAccessToCamera) {
-    switch (accessToCamera) {
-      case HasAccessToCamera.Yes:
-        return 'Yes';
-      case HasAccessToCamera.No:
-        return 'No';
-      case HasAccessToCamera.NotSure:
-        return 'Not sure';
-      default:
-        throw new Error(`Unexpected answer to computer question: ${accessToCamera}`);
-    }
   }
 }
