+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/individual-journey/pages/about-your-computer/about-your-computer.component.spec.ts
@@ -1,4 +1,4 @@
-import { HasAccessToCamera } from './../../../base-journey/participant-suitability.model';
+import { HasAccessToCamera } from '../../individual-suitability.model';
 import { IndividualJourneyStubs } from './../individual-base-component/individual-component-test-bed.spec';
 import { AboutYourComputerComponent } from './about-your-computer.component';
 import { IndividualJourneyComponentTestBed } from '../individual-base-component/individual-component-test-bed.spec';
