+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/individual-journey/services/individual-model-mapper.spec.ts
@@ -1,7 +1,7 @@
-import { SuitabilityAnswer, HasAccessToCamera, SelfTestAnswers } from '../../base-journey/participant-suitability.model';
+import { SuitabilityAnswer, SelfTestAnswers } from '../../base-journey/participant-suitability.model';
 import { HearingSuitabilityResponse, HearingSuitabilityAnswer } from '../../../services/clients/api-client';
 import { IndividualModelMapper, IndividualQuestionKeys } from './individual-model-mapper';
-import { IndividualSuitabilityModel } from '../individual-suitability.model';
+import { IndividualSuitabilityModel, HasAccessToCamera } from '../individual-suitability.model';
 import { MutableIndividualSuitabilityModel } from '../mutable-individual-suitability.model';
 import { SelfTestQuestionKeys } from '../../base-journey/services/participant-model-mapper';
